+++ TrabajoPractico3/Models/Cliente.cs
@@ -11,13 +11,13 @@ namespace TrabajoPractico3.Models
         private static int id = 0;
         private string nombre;
         private string direccion;
-        private int telefono;
+        private long telefono;
 
         public Cliente()
         {
         }
 
-        public Cliente(string nombre, string direccion, int telefono)
+        public Cliente(string nombre, string direccion, long telefono)
         {
             id++;
             this.nombre = nombre;
@@ -28,6 +28,6 @@ namespace TrabajoPractico3.Models
         public int Id { get => id; set => id = value; }
         public string Nombre { get => nombre; set => nombre = value; }
         public string Direccion { get => direccion; set => direccion = value; }
-        public int Telefono { get => telefono; set => telefono = value; }
+        public long Telefono { get => telefono; set => telefono = value; }
     }
 }
